requirements.txt
annotated-types==0.6.0
anyio==3.7.1
asgiref==3.7.2
backports.zoneinfo==0.2.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.5
defusedxml==0.8.0rc2
distro==1.8.0
Django==4.2.7
django-ranged-response==0.2.0
django-simple-captcha==0.5.20
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djoser==2.2.2
drf-yasg==1.21.7
exceptiongroup==1.2.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.4
inflection==0.5.1
oauthlib==3.2.2
openai==1.3.5
packaging==23.2
Pillow==10.1.0
pycparser==2.21
pydantic==2.5.2
pydantic-core==2.14.5
PyJWT==2.8.0
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rest-framework-simplejwt==0.0.2
sniffio==1.3.0
social-auth-app-django==5.4.0
social-auth-core==4.5.0
sqlparse==0.4.4
tqdm==4.66.1
typing-extensions==4.8.0
uritemplate==4.1.1
urllib3==2.1.0
